made health/mana integers added upgrades for health/mana

+++ src/bot.cpp
@@ -478,7 +478,7 @@ void _Bot::HandlePacket(_Buffer &Data) {
 
 					// No damage dealt
 					if((ActionResult.ActionUsed.GetTargetType() == TargetType::ENEMY || ActionResult.ActionUsed.GetTargetType() == TargetType::ENEMY_ALL)
-					   && ((ActionResult.Target.HasStat(StatType::HEALTH) && ActionResult.Target.Values[StatType::HEALTH].Float == 0.0f) || ActionResult.Target.HasStat(StatType::MISS))) {
+					   && ((ActionResult.Target.HasStat(StatType::HEALTH) && ActionResult.Target.Values[StatType::HEALTH].Integer == 0) || ActionResult.Target.HasStat(StatType::MISS))) {
 						ActionResult.Timeout = HUD_ACTIONRESULT_TIMEOUT_SHORT;
 						ActionResult.Speed = HUD_ACTIONRESULT_SPEED_SHORT;
 					}
@@ -498,10 +498,10 @@ void _Bot::HandlePacket(_Buffer &Data) {
 			HandleStatChange(Data, StatChange);
 		} break;
 		case PacketType::WORLD_HUD: {
-			Player->Health = Data.Read<float>();
-			Player->Mana = Data.Read<float>();
-			Player->MaxHealth = Data.Read<float>();
-			Player->MaxMana = Data.Read<float>();
+			Player->Health = Data.Read<int>();
+			Player->Mana = Data.Read<int>();
+			Player->MaxHealth = Data.Read<int>();
+			Player->MaxMana = Data.Read<int>();
 			Player->Experience = Data.Read<int32_t>();
 			Player->Gold = Data.Read<int32_t>();
 			double Clock = Data.Read<double>();
