+++ src/objects/battle.cpp
@@ -172,7 +172,7 @@ void _Battle::RenderActionResults(_ActionResult &ActionResult, double BlendFacto
 
 	// Draw damage dealt
 	glm::vec4 TextColor = COLOR_WHITE;
-	if(ActionResult.Target.HasStat(StatType::HEALTH) && ActionResult.Target.Values[StatType::HEALTH].Float > 0)
+	if(ActionResult.Target.HasStat(StatType::HEALTH) && ActionResult.Target.Values[StatType::HEALTH].Integer > 0)
 		TextColor = COLOR_GREEN;
 	else if(ActionResult.Target.HasStat(StatType::CRIT) && ActionResult.Target.Values[StatType::CRIT].Integer)
 		TextColor = COLOR_YELLOW;
@@ -183,7 +183,7 @@ void _Battle::RenderActionResults(_ActionResult &ActionResult, double BlendFacto
 	if(ActionResult.Target.HasStat(StatType::MISS))
 		Buffer << "miss";
 	else if(ActionResult.Target.HasStat(StatType::HEALTH))
-		Buffer << std::abs(ActionResult.Target.Values[StatType::HEALTH].Float);
+		Buffer << std::abs(ActionResult.Target.Values[StatType::HEALTH].Integer);
 	Assets.Fonts["hud_medium"]->DrawText(Buffer.str().c_str(), DrawPosition + glm::vec2(0, 7), TextColor, CENTER_BASELINE);
 }
